Add functionality to change grid size
To explain some of the deletions: there were some leftovers from
previous changes left in the code and they were only removed in this
commit.

Now the container has fixed dimensions and the repeat function of the
grid is what receives a variable. The create grid function now receives
a parameter with the dimensions for the grid. The "clear container"
function was created. It just resets the innerHTML of the container,
removing all the divs that were created previously.